try z as well as l routes

+++ src/schematic_from_netlist/cli.py
@@ -48,16 +48,21 @@ def produce_graph(db):
     junctions = router.insert_routing_junctions()
     db.insert_route_guide_buffers(junctions)
     db.dump_to_table("route_guide_insertion")
-    # db.geom2fig()
-    # gv.generate_layout_figures(phase="placed")
-    db.remove_multi_fanout_buffers()
-    db.geom2shape()
+    bypass_phase2 = True
+    if bypass_phase2:
+        db.remove_multi_fanout_buffers()
+        db.geom2shape()
+    else:
+        db.geom2fig()
+        gv.generate_layout_figures(phase="placed")
 
 
 def generate_schematic(db, output_dir: str):
     """Generate schematic info and produce LTSpice output."""
     db.schematic_db = GenSchematicData(db)
-    # db.schematic_db.generate_schematic()
+    bypass_phase2 = True
+    if not bypass_phase2:
+        db.schematic_db.generate_schematic()
 
     writer = LTSpiceWriter(db)
     writer.produce_schematic(output_dir)
